# Backend/technical-docs/api-contracts/openapi.yaml
...
openapi: 3.0.0
info:
  title: Farm E-Store API
  version: 0.0.0
...
paths:
  /api/v1/auth/request-otp:
    post:
      tags: [Auth]
      summary: Request OTP for login/registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mobile, user_type]
              properties:
                mobile:
                  type: string
                user_type:
                  type: string
                  enum: [Farmer, Agri-Retailer, Agent]
      responses:
        "200":
          description: OTP sent
...
  /api/v1/auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verify OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mobile, otp]
              properties:
                mobile:
                  type: string
                otp:
                  type: string
      responses:
        "200":
          description: OTP verified
...
  /api/admin/admin-login:
    post:
      tags: [Admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Admin logged in
...
  /api/user/login:
    post:
      tags: [Users]
      summary: User login (legacy email/password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: User logged in
  /api/user/add-user:
    post:
      tags: [Users]
      summary: Create user (Farmer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, mobile, user_type]
              properties:
                name:
                  type: string
                email:
                  type: string
                mobile:
                  type: string
                user_type:
                  type: string
      responses:
        "200":
          description: User created
  /api/user/me:
    get:
      tags: [Users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
    put:
      tags: [Users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
      responses:
        "200":
          description: User updated
  /api/user/get-user:
    get:
      tags: [Users]
      summary: Get all users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users list
  /api/user/get-user-by-id/{id}:
    get:
      tags: [Users]
      summary: Get user by id
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
  /api/user/update-user/{id}:
    put:
      tags: [Users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User updated
  /api/user/delete-user/{id}:
    delete:
      tags: [Users]
      summary: Delete user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
  /api/v1/addresses:
    post:
      tags: [Addresses]
      summary: Add address
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label, line1, city, state, postal_code]
              properties:
                label:
                  type: string
                tag:
                  type: string
                name:
                  type: string
                phone:
                  type: string
                line1:
                  type: string
                line2:
                  type: string
                city:
                  type: string
                state:
                  type: string
                postal_code:
                  type: string
                country:
                  type: string
                is_default:
                  type: boolean
      responses:
        "201":
          description: Address created
    get:
      tags: [Addresses]
      summary: List addresses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Address list
...
  /api/v1/cart:
    get:
      tags: [Cart]
      summary: Get cart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cart
    put:
      tags: [Cart]
      summary: Replace cart
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      product_id:
                        type: string
                      variant_id:
                        type: string
                      qty:
                        type: number
                      unit_price:
                        type: number
                      original_price:
                        type: number
                      commission:
                        type: number
      responses:
        "200":
          description: Cart updated
...
  /api/v1/coupons:
    post:
      tags: [Coupons]
      summary: Create coupon (Admin only)
      security:
        - adminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, discount_type, value, expires_at]
              properties:
                code:
                  type: string
                  description: Coupon code (will be uppercased)
                user_id:
                  type: string
                  description: Optional user ID (coupon specific to user)
                discount_type:
                  type: string
                  enum: [flat, percent]
                value:
                  type: number
                  description: Discount amount (flat) or percentage
                min_order:
                  type: number
                  description: Minimum order amount required
                expires_at:
                  type: string
                  format: date-time
                usage_limit:
                  type: number
                  description: Maximum number of times coupon can be used
      responses:
        "201":
          description: Coupon created
        "400":
          description: Invalid input
        "500":
          description: Server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT